api/requirements.txt
absl-py==1.4.0
addict==2.4.0
anyio==3.6.2
basicsr==1.4.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
facexlib==0.2.5
fastapi==0.93.0
filterpy==1.4.5
fonttools==4.39.0
future==0.18.3
gfpgan==1.3.8
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
h11==0.14.0
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
kiwisolver==1.4.4
lazy_loader==0.1
llvmlite==0.39.1
lmdb==1.4.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
networkx==3.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
PyWavelets==1.4.1
PyYAML==6.0
realesrgan==0.3.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tb-nightly==2.13.0a20230308
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tifffile==2023.2.28
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.3
yapf==0.32.0
zipp==3.15.0
